Moved subscribe into the prototype of the silo.
